refactored and started sqlshift in test
